+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.0.6] - 2023-06-18
+
+### Added
+
+- add `reveal in explorer` command to treeItem
+- add `📦 List folder's files in size-tree` to explorer context menu
+- add `sizeTree.useExcludeDefault` to settings
+
 ## [0.0.5] - 2023-06-08
 
 ### Fixed